fixed the working branch issue not committing my changes - users can now log in and comment and vote on posts - cant access comments or voting if not signed in. Cant log in if already logged in.
