+++ controllers/api/home-routes.js
@@ -1,0 +1,89 @@
+const router = require('express').Router();
+ 
+const sequelize = require('../config/connection');
+const { Post, User, Comment, Vote } = require('../models');
+ 
+ 
+// get all posts for homepage
+router.get('/', (req, res) => {
+   console.log('======================');
+   Post.findAll({
+     attributes: [
+       'id',
+       'post_url',
+       'title',
+       'created_at',
+       [sequelize.literal('(SELECT COUNT(*) FROM vote WHERE post.id = vote.post_id)'), 'vote_count']
+     ],
+     include: [
+       {
+         model: Comment,
+         attributes: ['id', 'comment_text', 'post_id', 'user_id', 'created_at'],
+         include: {
+           model: User,
+           attributes: ['username']
+         }
+       },
+       {
+         model: User,
+         attributes: ['username']
+       }
+     ]
+   })
+     .then(dbPostData => {
+       const posts = dbPostData.map(post => post.get({ plain: true }));
+        res.render('homepage', { posts, loggedIn: req.session.loggedIn });
+     })
+     .catch(err => {
+       console.log(err);
+       res.status(500).json(err);
+     });
+ });
+  router.get('/post/:id', (req, res) => {
+   Post.findOne({
+       where: {
+         id: req.params.id
+       },
+       attributes: [
+         'id',
+         'post_url',
+         'title',
+         'created_at',
+         [sequelize.literal('(SELECT COUNT(*) FROM vote WHERE post.id = vote.post_id)'), 'vote_count']
+       ],
+       include: [
+         {
+           model: Comment,
+           attributes: ['id', 'comment_text', 'post_id', 'user_id', 'created_at'],
+           include: {
+             model: User,
+             attributes: ['username']
+           }
+         },
+         {
+           model: User,
+           attributes: ['username']
+         }
+       ]
+     })
+     .then(dbPostData => {
+         const post = dbPostData.get({plain: true});
+       console.log(dbPostData.post)
+       res.render('single-post', { post, loggedIn: req.session.loggedIn });
+     })
+ 
+       .catch(err => {
+         console.log(err);
+         res.status(500).json(err);
+       });
+ })
+ 
+ router.get('/login', (req, res) => {
+   if (req.session.loggedIn) {
+     res.redirect('/');
+     return;
+   }
+    res.render('login');
+ });
+ 
+module.exports = router;
