+++ public/javascript/comment.js
@@ -1,13 +1,10 @@
 async function commentFormHandler(event) {
     event.preventDefault();
-  
-    const comment_text = document.querySelector('textarea[name="comment-body"]').value.trim();
-  
-    const post_id = window.location.toString().split('/')[
+     const comment_text = document.querySelector('textarea[name="comment-body"]').value.trim();
+     const post_id = window.location.toString().split('/')[
       window.location.toString().split('/').length - 1
     ];
-  
-    if (comment_text) {
+     if (comment_text) {
         const response = await fetch('/api/comments', {
           method: 'POST',
           body: JSON.stringify({
@@ -18,7 +15,7 @@ async function commentFormHandler(event) {
             'Content-Type': 'application/json'
           }
         });
-      
+     
         if (response.ok) {
           document.location.reload();
         } else {
@@ -26,5 +23,5 @@ async function commentFormHandler(event) {
         }
       }
   }
-  
-  document.querySelector('.comment-form').addEventListener('submit', commentFormHandler);
+   document.querySelector('.comment-form').addEventListener('submit', commentFormHandler);
+
